Factorize some coverage testutils functions
Those functions will be used by Rust and Python coverage all together.
